% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{all_fields_exist}
\alias{all_fields_exist}
\title{Check all fields provided exist in 'support_fields' table.}
\usage{
all_fields_exist(fields, con)
}
\arguments{
\item{fields}{A data.frame or childs. Existing fields in the object.}

\item{con}{A connection object to the relevant database.}
}
\description{
Check all fields provided exist in 'support_fields' table.
}
