% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{db_create}
\alias{db_create}
\title{Creates a bespoke SQLite database for the \href{https://bautheac.github.io/finRes/}{\pkg{finRes}}
  universe}
\usage{
db_create(path = NULL, n = 10L, verbose = TRUE)
}
\arguments{
\item{path}{a scalar chatacter vector. Specifies target directory for the database file.
Defaults to home directory.}

\item{n}{a scalar integer integer vector. Specifies the number of database tables for
create for historical data storage. This is a storage versus time trade-off paramater.
More tables occupy more space on disk but offer better read/write performance. Defaults to 10.}

\item{verbose}{a logical scalar vector. Should progression messages be printed? Defaults to TRUE.}
}
\description{
Creates a SQLite database bespoke to the
  \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} universe. In particular, the
  database is designed to store Bloomberg data retireved using the
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.
}
\details{
Creates a SQLite database bespoke for use within the
  \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite context. The database is created
  in the destination directory specified in \code{path}. Populate with \code{\link{db_store}}.
}
\examples{
\dontrun{db_create()}


}
\seealso{
The \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite,
  in particular the \href{https://github.com/bautheac/pullit/}{\pkg{pullit}},
  \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}},
  \href{https://github.com/bautheac/fewISOs/}{\pkg{fewISOs}},
  \href{https://github.com/bautheac/GICS/}{\pkg{GICS}} &
  \href{https://github.com/bautheac/factorem/}{\pkg{factorem}} packages.
}
