% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{db_delete}
\alias{db_delete}
\title{Deletes data from a \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite
  database.}
\usage{
db_delete(file = NULL, instrument = "all", book = "all",
  name = "all")
}
\arguments{
\item{file}{a scalar character vector. Specifies the path to the appropriate 'storethat.sqlite'
file.}

\item{instrument}{a scalar character vector. Specifies the financial instruments to get a
snapshot for. Must be one of 'all', equity', 'fund' or 'futures'.}

\item{book}{a scalar character vector. Instrument dependent; for a given instrument, specifies
the book for the snapshot; 'all' deletes all data for a given instrument.}

\item{name}{a scalar character vector. Instrument dependent; for a given instrument, specifies
the name for the snapshot; 'all' deletes all data for a given instrument and book.}
}
\description{
Deletes all or some of the data content of a
  \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite database.
}
\examples{
\dontrun{db_delete()}


}
\seealso{
The \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite,
  in particular the \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} &
  \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} packages.
}
