% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{db_names}
\alias{db_names}
\title{What's in that \code{storethat} SQLite database?}
\usage{
db_names(file = NULL, instrument = "all", type = "all")
}
\arguments{
\item{file}{a scalar chatacter vector. Specifies the path to the
appropriate 'storethat.sqlite' file.}

\item{instrument}{a scalar chatacter vector. Specifies the type of financial instrument(s) to update.
One of: 'all', futures' or 'equity'. Defaults to 'all' which updates the whole database.}

\item{type}{a scalar chatacter vector. Specifies the type of data to update for the selected type of
  financial instrument. Intrument specific:
\itemize{
  \item{'futures':
    \itemize{
      \item{'all': updates all futures data.}
      \item{'term structure'.}
      \item{'aggregate.}
      \item{'CFTC.}
    }
  }
  \item{'equity':
    \itemize{
      \item{'all': updates all equity data.}
      \item{'market'.}
      \item{'key stats'.}
      \item{'income statement'.}
      \item{'balance sheet'.}
      \item{'cash flow statement'.}
      \item{'ratios'.}
    }
  }
}}
}
\value{
A data.table with columns \code{ticker}, \code{instrument}, \code{data type}, \code{start} & \code{end}.
}
\description{
Gives a snapshot of the names (tickers) for which a particular \code{storethat} SQLite database
  has data records. For each name found, details the type(s) of data found as well as the corresponding period
 covered.
}
\examples{
\dontrun{db_names()}

}
