% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{db_snapshot}
\alias{db_snapshot}
\title{Snapshot of a \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite
  database.}
\usage{
db_snapshot(file = NULL, instrument, book = "all", name = "all")
}
\arguments{
\item{file}{a scalar character vector. Specifies the path to the appropriate 'storethat.sqlite'
file.}

\item{instrument}{a scalar character vector. Specifies the financial instruments to get a
snapshot for. Must be one of 'all', equity', 'fund' or 'futures'.}

\item{book}{a scalar character vector. Instrument dependent; for a given instrument, specifies
the book for the snapshot; 'all' snapshots all the books available for the given instrument.}

\item{name}{a scalar character vector. Instrument dependent; for a given instrument, specifies
the name for the snapshot; 'all' snapshots all the names available for the given instrument.}
}
\description{
Provides a snapshot of the data stored in an existing
  \href{https://github.com/bautheac/storethat/}{\pkg{storethat}} SQLite database.
}
\examples{
\dontrun{db_snapshot()}


}
\seealso{
The \href{https://bautheac.github.io/finRes/}{\pkg{finRes}} suite,
  in particular the \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} &
  \href{https://github.com/bautheac/BBGsymbols/}{\pkg{BBGsymbols}} packages.
}
