% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{db_store,EquityBook-method}
\alias{db_store,EquityBook-method}
\title{Store method for S4 objects of class \linkS4class{EquityBook}.}
\usage{
\S4method{db_store}{EquityBook}(object, file = NULL, verbose = TRUE)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{EquityBook} from the
\code{pullit} package.}

\item{file}{a scalar character vector. Specifies the path to the
appropriate 'storethat.sqlite' file.}
}
\description{
Store method for S4 objects of class \linkS4class{EquityBook}.
}
