% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.r
\docType{methods}
\name{db_store,FuturesTS-method}
\alias{db_store,FuturesTS-method}
\title{Store method for S4 objects of class \linkS4class{FuturesTS}.}
\usage{
\S4method{db_store}{FuturesTS}(object, file = NULL)
}
\arguments{
\item{object}{an S4 object of class \code{FuturesTS} from the
\code{pullit} package.}

\item{file}{a scalar character vector. Specifies the path to the
appropriate 'storethat.sqlite' file.}
}
\description{
Store method for S4 objects of class \linkS4class{FuturesTS}.
}
