% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r, R/methods.r
\docType{methods}
\name{db_store}
\alias{db_store}
\alias{db_store,FuturesTS-method}
\alias{db_store,FuturesTS}
\alias{db_store,FuturesAggregate-method}
\alias{db_store,FuturesAggregate}
\alias{db_store,FuturesSpot-method}
\alias{db_store,FuturesSpot}
\alias{db_store,FuturesCFTC-method}
\alias{db_store,FuturesCFTC}
\alias{db_store,FuturesInfo-method}
\alias{db_store,FuturesInfo}
\alias{db_store,EquityMarket-method}
\alias{db_store,EquityMarket}
\alias{db_store,EquityBook-method}
\alias{db_store,EquityBook}
\alias{db_store,EquityInfo-method}
\alias{db_store,EquityInfo}
\alias{db_store,FundMarket-method}
\alias{db_store,FundMarket}
\alias{db_store,FundInfo-method}
\alias{db_store,FundInfo}
\alias{db_store,IndexMarket-method}
\alias{db_store,IndexMarket}
\alias{db_store,IndexInfo-method}
\alias{db_store,IndexInfo}
\title{Stores financial data retrieved from Bloomberg with
  \href{https://github.com/bautheac/pullit/}{\pkg{pullit}}}
\usage{
db_store(object, file = NULL, verbose = TRUE)

\S4method{db_store}{FuturesTS}(object, file = NULL, verbose = TRUE)

\S4method{db_store}{FuturesAggregate}(object, file = NULL,
  verbose = TRUE)

\S4method{db_store}{FuturesSpot}(object, file = NULL, verbose = TRUE)

\S4method{db_store}{FuturesCFTC}(object, file = NULL, verbose = TRUE)

\S4method{db_store}{FuturesInfo}(object, file)

\S4method{db_store}{EquityMarket}(object, file = NULL, verbose = TRUE)

\S4method{db_store}{EquityBook}(object, file = NULL, verbose = TRUE)

\S4method{db_store}{EquityInfo}(object, file)

\S4method{db_store}{FundMarket}(object, file = NULL, verbose = TRUE)

\S4method{db_store}{FundInfo}(object, file)

\S4method{db_store}{IndexMarket}(object, file = NULL, verbose = TRUE)

\S4method{db_store}{IndexInfo}(object, file)
}
\arguments{
\item{object}{an S4 object of class \linkS4class{DataHistorical} or \linkS4class{DataInfo}
and respective childs.}

\item{file}{a scalar character vector. Specifies the path to the appropriate 'storethat.sqlite'
file.}

\item{verbose}{A logical scalar vector. Should progression messages be printed? Defaults to TRUE.}
}
\description{
Stores in a bespoke \href{https://github.com/bautheac/storethat/}{\pkg{storethat}}
  SQLite database financial data retrieved from Bloomberg
  using the \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package
  (\href{https://bautheac.github.io/finRes/}{\pkg{finRes}}) suite.
}
\seealso{
The \href{https://github.com/bautheac/pullit/}{\pkg{pullit}} package.
}
