% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{db_store}
\alias{db_store}
\title{Generic method for storing financial historical data retrieved from Bloomberg
  via the \code{pullit} package.}
\usage{
db_store(object, file = NULL)
}
\arguments{
\item{object}{an S4 object of class \code{BBGHitorical} from the
\code{pullit} package.}

\item{file}{a scalar character vector. Specifies the path to the
appropriate 'storethat.sqlite' file.}
}
\description{
Generic method for storing financial historical data retrieved from Bloomberg
  via the \code{pullit} package.
}
