% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{db_update}
\alias{db_update}
\title{Update a \code{storethat} SQLite database with fresh up to date historical data from Bloomberg}
\usage{
db_update(file = NULL, instrument = "all", type = "all",
  verbose = TRUE)
}
\arguments{
\item{file}{a scalar chatacter vector. Specifies the path to the
appropriate 'storethat.sqlite' file.}

\item{instrument}{a scalar chatacter vector. Specifies the type of financial instrument(s) to update.
One of: 'all', futures' or 'equity'. Defaults to 'all' which updates the whole database.}

\item{type}{a scalar chatacter vector. Specifies the type of data to update for the selected type of
  financial instrument. Intrument specific:
\itemize{
  \item{'futures':
    \itemize{
      \item{'all': updates all futures data.}
      \item{'term structure'.}
      \item{'aggregate.}
      \item{'CFTC.}
    }
  }
  \item{'equity':
    \itemize{
      \item{'all': updates all equity data.}
      \item{'market'.}
      \item{'key stats'.}
      \item{'income statement'.}
      \item{'balance sheet'.}
      \item{'cash flow statement'.}
      \item{'ratios'.}
    }
  }
}}

\item{verbose}{a logical scalar vector. Should progression messages be printed? Defaults to TRUE.}
}
\description{
Updates historical data for financial instruments that already exists in a \code{storethat}
  SQLite database but which data times series are outdated.
}
\examples{
\dontrun{db_update()}
}
\seealso{
The \code{pullit} package form the \code{finRes} suite.
}
