% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storethat-package.R
\docType{package}
\name{storethat-package}
\alias{storethat}
\alias{storethat-package}
\title{storethat: Store Bloomberg financial data for later use}
\description{
Stores Bloomberg financial data retrieved with the pullit package (finRes suite) in a SQLite
  database; a collection of helper functions provide easy and fast access.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bautheac.github.io/storethat}
  \item Report bugs at \url{https://github.com/bautheac/storethat/issues}
}

}
\author{
\strong{Maintainer}: Olivier Bauthéac \email{o.bautheac@gmail.com}

}
\keyword{internal}
