% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{update_cftc_tickers}
\alias{update_cftc_tickers}
\title{Updates the 'tickers_support_futures_ts' table with provided term structure tickers.}
\usage{
update_cftc_tickers(tickers, con)
}
\description{
Updates the 'tickers_support_futures_ts' table with provided term structure tickers.
}
