% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{update_data_cftc}
\alias{update_data_cftc}
\title{Updates the data table specified with the historical data dataset provided.}
\usage{
update_data_cftc(data, table_data, tickers, dates, con)
}
\arguments{
\item{data}{A data.frame or childs. Contains candidate data for \code{table_data}
update.}

\item{table_data}{A scalar character vector. Specifies the name of the table
to udate with \code{data}.}

\item{tickers}{A data.frame or childs. Tickers corresponding to \code{data}.}

\item{dates}{A data.frame or childs. Dates corresponding to \code{data}.}

\item{con}{A connection object to the relevant database.}
}
\description{
Updates the data table specified with the historical data dataset provided.
}
