% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{update_details}
\alias{update_details}
\title{Updates the data details table specified with the details dataset provided.}
\usage{
update_details(details, table_tickers, table_details, date_id, tickers,
  fields, con)
}
\description{
Updates the data details table specified with the details dataset provided.
}
