% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{update_fields}
\alias{update_fields}
\title{Updates 'support_fields' table with provided fields.}
\usage{
update_fields(fields, con)
}
\arguments{
\item{fields}{A data.frame or childs. Existing fields in the object.}

\item{con}{A connection object to the relevant database.}
}
\description{
Updates 'support_fields' table with provided fields.
}
