% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{update_tickers}
\alias{update_tickers}
\title{Updates the specified support tickers table with provided tickers.}
\usage{
update_tickers(tickers, table_tickers, con)
}
\description{
Updates the specified support tickers table with provided tickers.
}
