% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{db_create}
\alias{db_create}
\title{A bespoke SQLite database for the finRes universe}
\usage{
db_create(path = NULL, verbose = TRUE)
}
\arguments{
\item{path}{A scalar chatacter vector. Specifies target directory for the database file. Defaults to home directory.}

\item{verbose}{A logical scalar vector. Should progression messages be printed? Defaults to TRUE.}
}
\description{
Creates a SQLite database bespoke to the finRes universe. In particular, the resulting database
  is designed to store Bloomberg data retireved using the \code{pullit} package. Once stored, the data can
  be easily accessed and conveniently used for analysis with, but not limited to, other \code{finRes} packages,
  including \code{factorem}.
}
\details{
Creates a SQLite database in the destination directory. Tables include:
\itemize{
  \item{support_countries. World countries and corresponding ISO 3166-1 codes. Columns include:
    \itemize{
      \item{id: integer. ID for relational queries.}
      \item{symbol: character. Country alpha 2 ISO 3166-1 two-letter alphabetic code.}
      \item{name: character. Country name.}
      \item{capital: character. Country's capital.}
    }
  }
  \item{support_currencies World currencies and corresponding ISO 4217 codes. Columns include:
    \itemize{
      \item{id: integer. ID for relational queries.}
      \item{symbol: character. Country alpha 2 ISO 3166-1 two-letter alphabetic code.}
      \item{name: character. Currency name.}
      \item{country_id: integer. Issuer country; reference to column 'id' of table 'support_countries'.}
    }
  }
  \item{support_dates. Daily dates from 1970-01-01 to 2030-12-31:
    \itemize{
      \item{id: integer. ID for relational queries.}
      \item{date: date. }
    }
  }
  \item{support_exchanges. World market exchanges and corresponding ISO 10383 Market Identification Codes (MIC). Columns include:
    \itemize{
      \item{id: integer. ID for relational queries.}
      \item{symbol: character. ISO 10383 Market Identification Codes (MIC).}
      \item{name: character. Exchange name.}
      \item{country_id: integer. Country where the exchange sits; reference to column 'id' of table 'support_countries'.}
      \item{city: character. City where the exchange sits.}
    }
  }
  \item{support_futures_roll_types. Roll type Bloomberg symbols for futures Bloomberg tickers construction. Columns include:
    \itemize{
      \item{id: integer. ID for relational queries.}
      \item{symbol: character. Bloomberg roll adjustment symbol.}
      \item{name: character. Roll adjustment method name.}
    }
  }
  \item{support_futures_roll_adjustments. Roll adjustment Bloomberg symbols for futures Bloomberg tickers construction. Columns include:
    \itemize{
      \item{id: integer. ID for relational queries.}
      \item{symbol: character. Bloomberg roll type symbol.}
      \item{name: character. Roll type name.}
    }
  }
  \item{support_GICS. Global industry classification standard (GICS) dataset from this package. See ?GICS for details.}
  \item{tickers_futures. Active contract Bloomberg tickers for popular futures series. Columns include:
    \itemize{
      \item{id: integer. ID for relational queries.}
      \item{symbol: character. Active contract Bloomberg ticker.}
      \item{name: character. Futures series name.}
      \item{asset_class: character. Futures' underlying asset class.}
      \item{sector: character. Futures' underlying sector.}
      \item{subsector: character. Futures' underlying sector.}
      \item{TS_length: integer. length of the term structure for the corresponding futures series (# contracts).}
      \item{MIC_id: integer. Reference to column 'id' of table 'support_exchanges'. Identifies the exchanges where the futures series trades.}
      \item{currency_id: integer. Reference to column 'id' of table 'support_currencies'. Identifies the currency in which the futures series trades.}
      \item{FIGI: Financial instrument global identifier (FIGI). 12-character alpha-numerical code that serves for uniform unique global financial
        instrument identification. See \url{https://openfigi.com/} for details.}
    }
  }
  \item{tickers_equity. Equity Bloomberg tickers. Columns include:
    \itemize{
      \item{id: integer. ID for relational queries.}
      \item{symbol: character. Equity Bloomberg ticker.}
      \item{name: character. Corporation name.}
      \item{type: character. Security's type.}
      \item{MIC_id: integer. Reference to column 'id' of table 'support_exchanges'. Identifies the exchanges where the security trades.}
      \item{country_id: integer. Country where the corporation is incorporated; reference to column 'id' of table 'support_countries'.}
      \item{currency_id: integer. integer. Reference to column 'id' of table 'support_currencies'. Identifies the currency in which the security trades.}
      \item{GICS_subindustry_id: integer. GICS subindustry the corporation belongs to; reference to column 'subindustry_id' of table 'support_GICS'.}
      \item{FIGI: Financial instrument global identifier (FIGI). 12-character alpha-numerical code that serves for uniform unique global financial
        instrument identification. See \url{https://openfigi.com/} for details.}
    }
  }
  \item{tickers_support_futures_cftc. Bloomberg tickers for multiple CFTC position reports. Columns include:
    \itemize{
      \item{id: integer. ID for relational queries.}
      \item{symbol: character. Bloomberg ticker.}
      \item{active_contract_ticker_id: integer. Reference to column 'id' of table 'tickers_futures'.
        Identifies the futures series the Bloomberg ticker belongs to.}
      \item{format: character. CFTC report format.}
      \item{underlying: character. Underlying instrument.}
      \item{unit: character. Counting unit.}
      \item{participant: character. CFTC market participant category.}
      \item{position: character. Participant's position.}
    }
  }
  \item{tickers_support_futures_ts. Futures term structure Bloomberg tickers. Columns include:
    \itemize{
      \item{id: integer. ID for relational queries.}
      \item{symbol: character. Futures term structure Bloomberg ticker.}
      \item{active_contract_ticker_id: integer. Reference to column 'id' of table 'tickers_futures'.
        Identifies the futures series the Bloomberg ticker belongs to.}
      \item{position: integer. Identifies the term strucutre position for the corresponding futures series.}
      \item{roll_type_id: integer. Reference to column 'id' of table 'support_futures_roll_types'.
        Identifies the Bloomberg roll type used for the corresonding futures series.}
      \item{roll_days: integer. Roll days adjustment for the corresonding futures series.}
      \item{roll_months: integer. Roll months adjustment for the corresonding futures series.}
      \item{roll_adjustment_id: integer. Reference to column 'id' of table 'support_futures_roll_adjustments'.
        Identifies the Bloomberg roll adjustment for the corresonding futures series.}
    }
  }
  \item{data_futures_cftc. Stores CFTC position historical data retrieved from Bloomberg via the
    \code{pullit} package. Columns include:
    \itemize{
      \item{ticker_id: integer. Reference to column 'id' of table 'tickers_support_futures_cftc'.}
      \item{date_id: integer. Reference to column 'id' of table 'support_dates'.}
      \item{value: numeric. Corresponding value.}
    }
  }
  \item{data_futures_ts. Stores futures term strucutre market historical data retrieved from Bloomberg
    via the \code{pullit} package. Columns include:
    \itemize{
      \item{ticker_id: integer. Reference to column 'id' of table 'tickers_support_futures_ts'.}
      \item{date_id: integer. Reference to column 'id' of table 'support_dates'.}
      \item{The other columns are Bloomberg futures market data fields listed in the \code{\link[bbgsymbols]{fields}}
        datasets of the \code{bbgsymbols} package.}
    }
  }
  \item{data_futures_aggregate. Stores futures aggregate market historical data retrieved from Bloomberg.
    via the \code{pullit} package. Columns include:
    \itemize{
      \item{ticker_id: integer. Reference to column 'id' of table 'tickers_futures'.}
      \item{date_id: integer. Reference to column 'id' of table 'support_dates'.}
      \item{The other columns are Bloomberg futures aggregate data fields listed in the \code{\link[bbgsymbols]{fields}}
        datasets of the \code{bbgsymbols} package.}
    }
  }
  \item{data_equity_bs. Stores equity balance sheet historical data retrieved from Bloomberg.
    via the \code{pullit} package. Columns include:
    \itemize{
      \item{ticker_id: integer. Reference to column 'id' of table 'tickers_equity'.}
      \item{date_id: integer. Reference to column 'id' of table 'support_dates'.}
      \item{The other columns are Bloomberg equity balance sheet data fields listed in the \code{\link[bbgsymbols]{fields}}
        datasets of the \code{bbgsymbols} package.}
    }
  }
  \item{data_equity_cf. Stores equity cash flow statement historical data retrieved from Bloomberg
    via the \code{pullit} package. Columns include:
    \itemize{
      \item{ticker_id: integer. Reference to column 'id' of table 'tickers_equity'.}
      \item{date_id: integer. Reference to column 'id' of table 'support_dates'.}
      \item{The other columns are Bloomberg equity cash flow statement data fields listed in the \code{\link[bbgsymbols]{fields}}
        datasets of the \code{bbgsymbols} package.}
    }
  }
  \item{data_equity_is. Stores equity income statement statement historical data retrieved from Bloomberg
    via the \code{pullit} package. Columns include:
    \itemize{
      \item{ticker_id: integer. Reference to column 'id' of table 'tickers_equity'.}
      \item{date_id: integer. Reference to column 'id' of table 'support_dates'.}
      \item{The other columns are Bloomberg equity income statement data fields listed in the \code{\link[bbgsymbols]{fields}}
        datasets of the \code{bbgsymbols} package.}
    }
  }
  \item{data_equity_ks. Stores equity key stats historical data retrieved from Bloomberg
    via the \code{pullit} package. Columns include:
    \itemize{
      \item{ticker_id: integer. Reference to column 'id' of table 'tickers_equity'.}
      \item{date_id: integer. Reference to column 'id' of table 'support_dates'.}
      \item{The other columns are Bloomberg equity key stats data fields listed in the \code{\link[bbgsymbols]{fields}}
        datasets of the \code{bbgsymbols} package.}
    }
  }
  \item{data_equity_market. Stores equity market historical data retrieved from Bloomberg
    via the \code{pullit} package. Columns include:
    \itemize{
      \item{ticker_id: integer. Reference to column 'id' of table 'tickers_equity'.}
      \item{date_id: integer. Reference to column 'id' of table 'support_dates'.}
      \item{The other columns are Bloomberg equity market data fields listed in the \code{\link[bbgsymbols]{fields}}
        datasets of the \code{bbgsymbols} package.}
    }
  }
  \item{data_equity_ratios. Stores equity ratios historical data retrieved from Bloomberg
    via the \code{pullit} package. Columns include:
    \itemize{
      \item{ticker_id: integer. Reference to column 'id' of table 'tickers_equity'.}
      \item{date_id: integer. Reference to column 'id' of table 'support_dates'.}
      \item{The other columns are Bloomberg equity ratios data fields listed in the \code{\link[bbgsymbols]{fields}}
        datasets of the \code{bbgsymbols} package.}
    }
  }
}
}
\examples{
\dontrun{db_create()}

}
\seealso{
\code{bbgsymbols}, \code{fewISOs}, \code{pullit} and \code{factorem} packages form the \code{finRes} suite.
}
