% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storethat-package.R
\docType{package}
\name{storethat-package}
\alias{storethat}
\alias{storethat-package}
\title{storethat: Store Bloomberg financial data for later use}
\description{
Stores Bloomberg financial data retrieved with the pullit package (finRes suite) in a SQLite
  database; a collection of helper functions provide easy and fast access.
}
\author{
\strong{Maintainer}: Olivier Bautheac \email{o.bautheac@gmail.com}

}
\keyword{internal}
