% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{update_data}
\alias{update_data}
\title{Updates the data table specified with the historical data dataset provided.}
\usage{
update_data(data, table_data, tickers, fields, dates, con)
}
\description{
Updates the data table specified with the historical data dataset provided.
}
