% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{update_data_cftc}
\alias{update_data_cftc}
\title{Updates the data table specified with the historical data dataset provided.}
\usage{
update_data_cftc(data, table_data, tickers, dates, con)
}
\description{
Updates the data table specified with the historical data dataset provided.
}
