% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{update_term_structure_tickers}
\alias{update_term_structure_tickers}
\title{Updates the 'tickers_support_futures_ts' table with provided term structure tickers.}
\usage{
update_term_structure_tickers(tickers, con)
}
\arguments{
\item{tickers}{A character vector. Existing tickers in the object.}

\item{con}{A connection object to the relevant database.}
}
\description{
Updates the 'tickers_support_futures_ts' table with provided term structure tickers.
}
