% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.r
\name{update_tickers}
\alias{update_tickers}
\title{Updates the specified support tickers table with provided tickers.}
\usage{
update_tickers(tickers, table_tickers, con)
}
\arguments{
\item{tickers}{A character vector. Existing tickers in the object.}

\item{table_tickers}{A scalar character vector. Specifies the table name
where comparison tickers sit.}

\item{con}{A connection object to the relevant database.}
}
\description{
Updates the specified support tickers table with provided tickers.
}
